pyproject.toml
[project]
name = "rsac-relatorios-risco"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "openpyxl",
    "python-dotenv",
    "loguru",
    "requests",
    "msal",
    "jarbis_external_client",
    "pywinauto",
    "selenium",
    "pyautogui",
    "pyperclip",
    "opencv-python",
    "psutil",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
